+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Substrings of the form `$name` or `${name}` are replaced by the value of environment variable name.
 - Send custom User-Agent with client version to registry API
 - Add OS, platform and python version to API client user agent
+- Provide list of managed components to ESP-IDF build system
 
 ### Changed